+++ db/migrate/20250613095029_add_aliases_to_memory_entities.rb
@@ -1,0 +1,5 @@
+class AddAliasesToMemoryEntities < ActiveRecord::Migration[8.0]
+  def change
+    add_column :memory_entities, :aliases, :text
+  end
+end